LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_navigation_download_block_downloads.sub.tentative-expected.txt

CONSOLE MESSAGE: Not allowed to download due to sandboxing


PASS Navigation resulted download in sandbox is blocked.
PASS Navigation resulted download in sandbox from <object> is blocked.
